add interfaces for State in quiz and question editors
