+++ apps/geek-platform/src/app/components/question-editor/question-editor.component.scss
@@ -7,5 +7,5 @@
 }
 
 .container-list.cdk-drop-list-dragging .list-item:not(.cdk-drag-placeholder) {
-  transition: transform 350ms ease-in-out;
+  transition: var(--drag-animating-transition);
 }
