+++ apps/geek-platform/src/app/components/question-editor/question-editor.interfaces.ts
@@ -1,0 +1,7 @@
+import { QuestionDto } from '@geek-platform/api-interfaces';
+
+export interface State {
+  question: QuestionDto;
+  isVisibleAddButton: boolean;
+  isVisibleDeleteButton: boolean;
+}
