+++ apps/geek-platform/src/app/components/quiz-editor/quiz-editor.component.scss
@@ -8,5 +8,5 @@
 }
 
 .container-list.cdk-drop-list-dragging .list-item:not(.cdk-drag-placeholder) {
-  transition: transform 350ms ease-in-out;
+  transition: var(--drag-animating-transition);
 }
