+++ apps/geek-platform/src/app/components/quiz-editor/quiz-editor.interfaces.ts
@@ -1,0 +1,11 @@
+import { QuestionDto } from '@geek-platform/api-interfaces';
+
+export interface State {
+  isVisibleAddButton: boolean;
+  isVisibleDeleteButton: boolean;
+  quiz: {
+    _id?: string;
+    name: string;
+    questions: Partial<QuestionDto>[];
+  };
+}
