+++ apps/geek-platform/src/styles.scss
@@ -1,38 +1,3 @@
-
-// Custom Theming for Angular Material
-// For more information: https://material.angular.io/guide/theming
-@import '~@angular/material/theming';
-// Plus imports for other components in your app.
-
-// Include the common styles for Angular Material. We include this here so that you only
-// have to load a single css file for Angular Material in your app.
-// Be sure that you only ever include this mixin once!
-@include mat-core();
-
-// Define the palettes for your theme using the Material Design palettes available in palette.scss
-// (imported above). For each palette, you can optionally specify a default, lighter, and darker
-// hue. Available color palettes: https://material.io/design/color/
-$geek-platform-primary: mat-palette($mat-indigo);
-$geek-platform-accent: mat-palette($mat-pink, A200, A100, A400);
-
-// The warn palette is optional (defaults to red).
-$geek-platform-warn: mat-palette($mat-red);
-
-// Create the theme object. A theme consists of configurations for individual
-// theming systems such as "color" or "typography".
-$geek-platform-theme: mat-light-theme((
-  color: (
-    primary: $geek-platform-primary,
-    accent: $geek-platform-accent,
-    warn: $geek-platform-warn,
-  )
-));
-
-// Include theme styles for core and each component used in your app.
-// Alternatively, you can import and @include the theme mixins for each component
-// that you are using.
-@include angular-material-theme($geek-platform-theme);
-
 :root {
   --light: #eeeeee;
   --gray0: #e4e4e4;
@@ -56,6 +21,8 @@ $geek-platform-theme: mat-light-theme((
   --nav-header-space: 80px;
   --container-width: 700px;
   --nav-header-height: calc(var(--space-atom) * 8 + 1rem);
+
+  --drag-animating-transition: transform var(--transition) ease-in-out;
 }
 
 .light {
@@ -142,5 +109,5 @@ $spaces: 1 2 3 4 5;
 }
 
 .cdk-drag-animating {
-  transition: transform 350ms ease-in-out;
+  transition: var(--drag-animating-transition);
 }
